webapps/requirements.txt
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.6
django-storages==1.1.8
django-toolbelt==0.0.1
gunicorn==19.3.0
Pillow==3.0.0
psycopg2==2.6.1
static3==0.6.1
wheel==0.24.0
